Refactor plotting code
