+++ src/dkpy/__init__.py
@@ -3,5 +3,6 @@
 from .controller_synthesis import *
 from .dk_iteration import *
 from .fit_transfer_functions import *
+from .plotting import *
 from .structured_singular_value import *
 from .utilities import *
