+++ src/dkpy/dk_iteration.py
@@ -14,7 +14,6 @@ from typing import Any, Callable, Dict, List, Tuple, Union, Optional
 import control
 import numpy as np
 import scipy.linalg
-from matplotlib import pyplot as plt
 
 from . import (
     controller_synthesis,
@@ -65,139 +64,6 @@ class DScaleFitInfo:
         self.D_fit_omega = D_fit_omega
         self.D_fit = D_fit
         self.block_structure = block_structure
-
-    def plot_mu(
-        self,
-        ax: Optional[plt.Axes] = None,
-        plot_kw: Optional[Dict[str, Any]] = None,
-        hide: Optional[str] = None,
-    ) -> Tuple[plt.Figure, plt.Axes]:
-        """Plot mu.
-
-        Parameters
-        ----------
-        ax : Optional[plt.Axes]
-            Matplotlib axes to use.
-        plot_kw : Optional[Dict[str, Any]]
-            Keyword arguments for :func:`plt.Axes.semilogx`.
-        hide : Optional[str]
-            Set to ``'mu_omega'`` or ``'mu_fit_omega'`` to hide either one of
-            those lines.
-
-        Returns
-        -------
-        Tuple[plt.Figure, plt.Axes]
-            Matplotlib :class:`plt.Figure` and :class:`plt.Axes` objects.
-        """
-        # Create figure if not provided
-        if ax is None:
-            fig, ax = plt.subplots()
-        else:
-            fig = ax.get_figure()
-        # Set label
-        label = plot_kw.pop("label", "mu")
-        label_mu_omega = label + ""
-        label_mu_fit_omega = label + "_fit"
-        # Clear line styles
-        _ = plot_kw.pop("ls", None)
-        _ = plot_kw.pop("linestyle", None)
-        # Plot mu
-        if hide != "mu_omega":
-            ax.semilogx(
-                self.omega,
-                self.mu_omega,
-                label=label_mu_omega,
-                ls="--",
-                **plot_kw,
-            )
-        if hide != "mu_fit_omega":
-            ax.semilogx(
-                self.omega,
-                self.mu_fit_omega,
-                label=label_mu_fit_omega,
-                **plot_kw,
-            )
-        # Set axis labels
-        ax.set_xlabel(r"$\omega$ (rad/s)")
-        ax.set_ylabel(r"$\mu(\omega)$")
-        ax.grid(linestyle="--")
-        ax.legend(loc="lower left")
-        # Return figure and axes
-        return fig, ax
-
-    def plot_D(
-        self,
-        ax: Optional[np.ndarray] = None,
-        plot_kw: Optional[Dict[str, Any]] = None,
-        hide: Optional[str] = None,
-    ) -> Tuple[plt.Figure, np.ndarray]:
-        """Plot D.
-
-        Parameters
-        ----------
-        ax : Optional[np.ndarray]
-            Array of Matplotlib axes to use.
-        plot_kw : Optional[Dict[str, Any]]
-            Keyword arguments for :func:`plt.Axes.semilogx`.
-        hide : Optional[str]
-            Set to ``'D_omega'`` or ``'D_fit_omega'`` to hide either one of
-            those lines.
-
-        Returns
-        -------
-        Tuple[plt.Figure, np.ndarray]
-            Matplotlib :class:`plt.Figure` object and two-dimensional array of
-            :class:`plt.Axes` objects.
-        """
-        mask = fit_transfer_functions._mask_from_block_structure(self.block_structure)
-        # Create figure if not provided
-        if ax is None:
-            fig, ax = plt.subplots(
-                mask.shape[0],
-                mask.shape[1],
-                constrained_layout=True,
-            )
-        else:
-            fig = ax[0, 0].get_figure()
-        # Set label
-        label = plot_kw.pop("label", "D")
-        label_D_omega = label + ""
-        label_D_fit_omega = label + "_fit"
-        # Clear line styles
-        _ = plot_kw.pop("ls", None)
-        _ = plot_kw.pop("linestyle", None)
-        # Plot D
-        mag_D_omega = np.abs(self.D_omega)
-        mag_D_fit_omega = np.abs(self.D_fit_omega)
-        for i in range(ax.shape[0]):
-            for j in range(ax.shape[1]):
-                if mask[i, j] != 0:
-                    dB_omega = 20 * np.log10(mag_D_omega[i, j, :])
-                    dB_fit_omega = 20 * np.log10(mag_D_fit_omega[i, j, :])
-                    if hide != "D_omega":
-                        ax[i, j].semilogx(
-                            self.omega,
-                            dB_omega,
-                            label=label_D_omega,
-                            ls="--",
-                            **plot_kw,
-                        )
-                    if hide != "D_fit_omega":
-                        ax[i, j].semilogx(
-                            self.omega,
-                            dB_fit_omega,
-                            label=label_D_fit_omega,
-                            **plot_kw,
-                        )
-                    # Set axis labels
-                    ax[i, j].set_xlabel(r"$\omega$ (rad/s)")
-                    ax[i, j].set_ylabel(rf"$D_{i}{j}(\omega) (dB)$")
-                    ax[i, j].grid(linestyle="--")
-                else:
-                    ax[i, j].axis("off")
-        fig.legend(handles=ax[0, 0].get_lines(), loc="lower left")
-        # Return figure and axes
-        return fig, ax
 
     @classmethod
     def create_from_fit(
